Supprime et remplace l'ENUM TypesContratDeTravailDuree par l'ENUM TypesContrat
