+++ CHANGELOG.md
@@ -4952,8 +4952,8 @@ statut_marital == TypesStatutMarital.celibataire
 Sinon, importez-le dans votre formule avec l'attribut `possible_values` de la variable qui le calcule :
 
 ```py
-TypesContratDeTravailDuree = contrat_de_travail_type.possible_values
-contrat_travail == TypesContratDeTravailDuree.cdi
+TypesContrat = contrat_de_travail_type.possible_values
+contrat_travail == TypesContrat.cdi
 ```
 #### Effets sur la Web API
